test: Fix building dma_buf test on GCC 11
 In file included from dma_buf.cpp:13:
 dma_buf.cpp: In function ‘void VerifyTooLargeIndexFails(int)’:
 ioctl.h:38:41: warning: unsigned conversion from ‘int’ to ‘uint8_t’ {aka ‘unsigned char’} changes value from ‘256’ to ‘0’ [-Woverflow]
    38 | #define TENSTORRENT_MAX_DMA_BUFS        256
       |                                         ^~~
 dma_buf.cpp:75:60: note: in expansion of macro ‘TENSTORRENT_MAX_DMA_BUFS’
    75 |         auto buf_max = AllocateDmaBuf(dev_fd, page_size(), TENSTORRENT_MAX_DMA_BUFS);
       |

+++ test/dma_buf.cpp
@@ -34,7 +34,7 @@ std::size_t MaxDmaBufSize(int dev_fd)
 }
 
 std::variant<tenstorrent_allocate_dma_buf_out, int>
-AllocateDmaBuf(int dev_fd, std::uint32_t size, std::uint8_t index)
+AllocateDmaBuf(int dev_fd, std::uint32_t size, std::uint32_t index)
 {
     tenstorrent_allocate_dma_buf allocate_dma_buf;
     zero(&allocate_dma_buf);
